domain/src/main/java/ru/fastdelivery/domain/delivery/shipment/Shipment.java
package ru.fastdelivery.domain.delivery.shipment;

import ru.fastdelivery.domain.common.currency.Currency;
import ru.fastdelivery.domain.common.packproperties.Volume;
import ru.fastdelivery.domain.common.packproperties.weight.Weight;
import ru.fastdelivery.domain.delivery.pack.Pack;

import java.util.List;

/**
 * @param packages упаковки в грузе
 * @param currency валюта объявленная для груза
 */
public record Shipment(
        List<Pack> packages,
        Currency currency
) {
    public Weight weightAllPackages() {
        return packages.stream()
                .map(Pack::weight)
                .reduce(Weight.zero(), Weight::add);
    }

    public Volume volumeAllPackages(){
        return packages.stream()
                .map(pkg -> new Volume(pkg.length(), pkg.width(), pkg.height()))
                .reduce(Volume.zero(), Volume::add);
    }
}
